fix(Manipulators): remove manipulators by value instead of index
BREAKING CHANGE: manipulators array are now actually modified by removing
the manipulator from the list of manipulators, and not just replacing them
by null value for a certain index. This affects the resulting size of the
manipulator list, and users should not rely on an index that needed to
remain constant.